+++ packages/standard_repositories/lib/standard_repositories.dart
@@ -1,6 +1,6 @@
 /// A Very Good Project created by Very Good CLI.
 library;
 
-export 'src/fetcher_repository.dart';
+export 'src/multi_repository.dart';
 export 'src/repository_cache.dart';
 export 'src/standard_repositories.dart';