+++ packages/standard_repositories/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.1-dev.3
+
+- feat: add multi repository
+- feat: add repository groups
+- breaking: changed many of the functions calls
+
 ## 0.0.1-dev.2
 
 - feat: add repository caching